pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-harness"
version = "0.1.0"
description = "Minimal AI agent harness for live demos"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Eric Hansen" }]
dependencies = [
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pre-commit",
    "pytest",
]

[project.scripts]
agent-harness = "agent_harness.agent:main"
mcp-server = "mcp_server.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
